+++ demo3d.py
@@ -1,5 +1,6 @@
 import cv2
 import matplotlib.pyplot as plt
+import numpy as np
 import torch
 from PIL import Image
 from torchvision import transforms
@@ -11,8 +12,8 @@ from models.blazenet_model_3d import Pose3dModel
 config = {
     "model": {
         "n_keypoints": 21,
-        "model_file": "exp/model_1.pth",
-        "device": "cuda",
+        "model_file": "blazenet3d_combined_v1.pth",
+        "device": "cpu",
         "raw_image_size": 256,
         "model_img_size": 128,
     }
@@ -26,6 +27,18 @@ model.load_state_dict(
 image_transform = transforms.Compose(
     [transforms.Resize(config["model"]["raw_image_size"]), transforms.ToTensor(),]
 )
+
+# https://itectec.com/matlab/matlab-how-to-calculate-roll-pitch-and-yaw-from-xyz-coordinates-of-3-planar-points/
+def compute_hand_orientation(p1, p2, p3):
+    p1, p2, p3 = p1[:3], p2[:3], p3[:3]
+    x = (p1 + p2) / 2 - p3
+    v1, v2 = p2 - p1, p3 - p1
+    z = np.cross(v1, v2)
+    z = z / np.linalg.norm(z)
+    x = x / np.linalg.norm(x)
+    y = np.cross(z, x)
+    return x, y, z
+
 
 # define a video capture object
 vid = cv2.VideoCapture(0)
@@ -51,12 +64,30 @@ with torch.no_grad():
         heatmaps_pred = pred[0].cpu().numpy()[0]
         kpt_3d_pred = pred[1].cpu().numpy()[0]
 
+        p1 = kpt_3d_pred[5]
+        p2 = kpt_3d_pred[17]
+        p3 = kpt_3d_pred[0]
+        x, y, z = compute_hand_orientation(p1, p2, p3)
+        p4 = kpt_3d_pred[9]
+        p4x = p4 + x * 0.1
+        p4y = p4 + y * 0.1
+        p4z = p4 + z * 0.1
+
         kpt_2d_pred = heatmaps_to_coordinates(heatmaps_pred, config["model"]["model_img_size"])
         kpt_2d_pred[:, 0] = kpt_2d_pred[:, 0] * im_width
         kpt_2d_pred[:, 1] = kpt_2d_pred[:, 1] * im_height
         skeleton_overlay = draw_2d_skeleton(frame, kpt_2d_pred)
         ax.clear()
         draw_3d_skeleton_on_ax(kpt_3d_pred, ax)
+        ax.plot(
+            [p4[0], p4x[0]], [p4[1], p4x[1]], [p4[2], p4x[2]], zdir="z", c="red",
+        )
+        ax.plot(
+            [p4[0], p4y[0]], [p4[1], p4y[1]], [p4[2], p4y[2]], zdir="z", c="green",
+        )
+        ax.plot(
+            [p4[0], p4z[0]], [p4[1], p4z[1]], [p4[2], p4z[2]], zdir="z", c="blue",
+        )
 
         skeleton_overlay = cv2.cvtColor(skeleton_overlay, cv2.COLOR_RGB2BGR)
         cv2.imshow("frame", skeleton_overlay)
